Feat : 예매 취소에 Redis 확인 추가
예매취소에 Redis LS가 있으면 확인하는 기능 추가
현재 Redis에 저장된 LS 전체 키 조회 추가
ADMIN만 확인하도록 User 확인 추가

Related to: #64

+++ src/main/java/com/example/neticket/reservation/service/ReservationService.java
@@ -12,7 +12,11 @@ import com.example.neticket.reservation.entity.Reservation;
 import com.example.neticket.reservation.repository.RedisRepository;
 import com.example.neticket.reservation.repository.ReservationRepository;
 import com.example.neticket.user.entity.User;
+import com.example.neticket.user.entity.UserRoleEnum;
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -30,9 +34,7 @@ public class ReservationService {
   // 1. 예매중 페이지에서 공연정보 조회
   @Transactional(readOnly = true)
   public DetailEventResponseDto verifyReservation(Long ticketInfoId) {
-    return ticketInfoRepository.findById(ticketInfoId)
-        .map(ticketInfo -> new DetailEventResponseDto(ticketInfo.getEvent()))
-        .orElseThrow(() -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION));
+    return new DetailEventResponseDto(checkTicketInfoById(ticketInfoId).getEvent());
   }
 
   // 2. 예매하기
@@ -63,8 +65,7 @@ public class ReservationService {
 
   // 2-2. 캐시 없으면 DB로 좌석수 변경
   private void decrementLeftSeatInDB(ReservationRequestDto dto) {
-    TicketInfo ticketInfo = ticketInfoRepository.findById(dto.getTicketInfoId())
-        .orElseThrow(() -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION));
+    TicketInfo ticketInfo = checkTicketInfoById(dto.getTicketInfoId());
     if (!ticketInfo.isAvailable()) {
       throw new CustomException(ExceptionType.RESERVATION_UNAVAILABLE_EXCEPTION);
     }
@@ -79,28 +80,18 @@ public class ReservationService {
   // 3. 예매완료
   @Transactional(readOnly = true)
   public ReservationResponseDto reservationComplete(Long resvId, User user) {
-    Reservation reservation = reservationRepository.findById(resvId).orElseThrow(
-        () -> new CustomException(ExceptionType.NOT_FOUND_RESERVATION_EXCEPTION)
-    );
-    TicketInfo ticketInfo = ticketInfoRepository.findById(reservation.getTicketInfoId()).orElseThrow(
-        () -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION)
-    );
-
+    Reservation reservation = checkReservationById(resvId);
     checkReservationUser(reservation, user);
-
+    TicketInfo ticketInfo = checkTicketInfoById(reservation.getTicketInfoId());
     return new ReservationResponseDto(reservation, ticketInfo);
   }
 
-  // 4. 예매취소 여기에도 캐시 처리 해야
+  // 4. 예매취소 여기에도 캐시 처리 함
   @Transactional
   public void deleteReservation(Long resvId, User user) {
-    Reservation reservation = reservationRepository.findById(resvId).orElseThrow(
-        () -> new CustomException(ExceptionType.NOT_FOUND_RESERVATION_EXCEPTION)
-    );
+    Reservation reservation = checkReservationById(resvId);
     checkReservationUser(reservation, user);
-    TicketInfo ticketInfo = ticketInfoRepository.findById(reservation.getTicketInfoId()).orElseThrow(
-        () -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION)
-    );
+    TicketInfo ticketInfo = checkTicketInfoById(reservation.getTicketInfoId());
 
 //    공연날이 오늘이거나 오늘보다 이전이면 예매 취소 불가능
     LocalDate eventDay = LocalDate.from(ticketInfo.getEvent().getDate());
@@ -108,35 +99,66 @@ public class ReservationService {
       throw new CustomException(ExceptionType.CANCEL_DEADLINE_PASSED_EXCEPTION);
     }
 
-    // 예매 취소 후 좌석 수 업데이트
-    ticketInfo.plusSeats(reservation.getCount());
+    Integer leftSeats = redisRepository.findLeftSeatsFromRedis(ticketInfo.getId());
+    if (leftSeats == null) {
+      //    캐시가 없으면 DB를 통해 남은 좌석 수에 추가
+      ticketInfo.plusSeats(reservation.getCount());
+    } else {
+      //    캐시가 있으면 redis에서 남은 좌석 수에 추가
+      redisRepository.incrementLeftSeatInRedis(ticketInfo.getId(), reservation.getCount());
+    }
     reservationRepository.delete(reservation);
-
   }
 
-//  3-1, 4-1. 예매 기록의 사용자와 현재 토큰상의 사용자 일치 여부 판별 메서드
+  //  3-1. 예매 기록의 사용자와 현재 토큰상의 사용자 일치 여부 판별 메서드
   private void checkReservationUser(Reservation reservation, User user) {
     if (!reservation.getUser().getId().equals(user.getId())) {
       throw new CustomException(ExceptionType.USER_RESERVATION_NOT_MATCHING_EXCEPTION);
     }
   }
+  //  3-2. 예매 ID로 Reservation 확인
+  private Reservation checkReservationById(Long reservationId) {
+    return reservationRepository.findById(reservationId).orElseThrow(
+        () -> new CustomException(ExceptionType.NOT_FOUND_RESERVATION_EXCEPTION)
+    );
+  }
+  //  3-3. ticketInfoId로 TicketInfo 확인
+  private TicketInfo checkTicketInfoById(Long ticketInfoId) {
+    return ticketInfoRepository.findById(ticketInfoId).orElseThrow(
+        () -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION)
+    );
+  }
 
-  //  5. ADMIN. DB에서 남은 좌석수만 가져와서 Redis에 (key-value)형태로 저장
+  //  5. ADMIN. DB에서 남은 좌석수만 가져와서 Redis에 (key-value)형태로 저장.
   @Transactional
-  public MessageResponseDto saveLeftSeatsInRedis(Long ticketInfoId) {
-    TicketInfo ticketInfo = ticketInfoRepository.findById(ticketInfoId)
-        .orElseThrow(() -> new CustomException(ExceptionType.NOT_FOUND_TICKET_INFO_EXCEPTION));
+  public MessageResponseDto saveLeftSeatsInRedis(Long ticketInfoId, User user) {
+    checkAdmin(user);
+    TicketInfo ticketInfo = checkTicketInfoById(ticketInfoId);
     redisRepository.saveTicketInfoToRedis(ticketInfo);
     return new MessageResponseDto(HttpStatus.CREATED, "redis에 성공적으로 저장되었습니다.");
   }
 
+  //  5-1. user role이 admin인지 체크
+  private void checkAdmin(User user) {
+    if (!user.getRole().equals(UserRoleEnum.ADMIN)) {
+      throw new CustomException(ExceptionType.USER_UNAUTHORIZED_EXCEPTION);
+    }
+  }
+
   //  6. ADMIN. 해당하는 공연의 남은 좌석수 Redis에서 삭제(삭제되기전 모든 캐시 DB에 반영)
   @Transactional
-  public MessageResponseDto deleteLeftSeatsFromRedis(Long ticketInfoId) {
+  public MessageResponseDto deleteLeftSeatsFromRedis(Long ticketInfoId, User user) {
+    checkAdmin(user);
     redisRepository.deleteLeftSeatsInRedis(ticketInfoId);
     return new MessageResponseDto(HttpStatus.OK, "redis에서 캐시가 성공적으로 삭제되었습니다.");
   }
 
+//  7. ADMIN. 현재 Redis에 등록된 모든 LeftSeats의 key를 리스트로 반환
+  @Transactional(readOnly = true)
+  public List<String> findAllLeftSeatsKeysInRedis(User user) {
+    checkAdmin(user);
+    return new ArrayList<>(redisRepository.findAllLeftSeatsKeysInRedis());
+  }
 
 
 }
